session: move some tests to integration
It's a good practice to have tests which use only the public API
in the integration tests, to ensure that the public API is usable
and works as expected. Move some session-related tests to the
integration tests.
