+++ scylla-rust-wrapper/tests/integration/main.rs
@@ -1,2 +1,3 @@
 mod consistency;
+mod session;
 mod utils;
