+++ scylla-rust-wrapper/tests/integration/utils.rs
@@ -25,6 +25,30 @@ pub(crate) fn setup_tracing() {
         .try_init();
 }
 
+unsafe fn write_str_to_c(s: &str, c_str: *mut *const c_char, c_strlen: *mut size_t) {
+    unsafe {
+        *c_str = s.as_ptr() as *const c_char;
+        *c_strlen = s.len() as u64;
+    }
+}
+
+pub(crate) fn str_to_c_str_n(s: &str) -> (*const c_char, size_t) {
+    let mut c_str = std::ptr::null();
+    let mut c_strlen = size_t::default();
+
+    // SAFETY: The pointers that are passed to `write_str_to_c` are compile-checked references.
+    unsafe { write_str_to_c(s, &mut c_str, &mut c_strlen) };
+
+    (c_str, c_strlen)
+}
+
+macro_rules! make_c_str {
+    ($str:literal) => {
+        concat!($str, "\0").as_ptr() as *const c_char
+    };
+}
+pub(crate) use make_c_str;
+
 pub(crate) async fn test_with_3_node_dry_mode_cluster<I, F>(
     initial_request_rules: impl Fn() -> I,
     test: F,
@@ -58,6 +82,7 @@ where
     running_proxy.finish().await
 }
 
+#[track_caller]
 pub(crate) fn assert_cass_error_eq(errcode1: CassError, errcode2: CassError) {
     unsafe {
         assert_eq!(
@@ -128,6 +153,33 @@ pub(crate) fn handshake_rules() -> impl IntoIterator<Item = RequestRule> {
             })),
         ),
     ]
+}
+
+// As these are very generic, they should be put last in the rules Vec.
+pub(crate) fn generic_drop_queries_rules() -> impl IntoIterator<Item = RequestRule> {
+    [RequestRule(
+        Condition::RequestOpcode(RequestOpcode::Query),
+        // We won't respond to any queries (including metadata fetch),
+        // but the driver will manage to continue with dummy metadata.
+        RequestReaction::forge().server_error(),
+    )]
+}
+
+/// A set of rules that are needed to finish session initialization.
+// They are used in tests that require a session to be connected.
+// All connections are successfully negotiated.
+// All requests are replied with a server error.
+pub(crate) fn mock_init_rules() -> impl IntoIterator<Item = RequestRule> {
+    handshake_rules()
+        .into_iter()
+        .chain(std::iter::once(RequestRule(
+            Condition::RequestOpcode(RequestOpcode::Query)
+                .or(Condition::RequestOpcode(RequestOpcode::Prepare))
+                .or(Condition::RequestOpcode(RequestOpcode::Batch)),
+            // We won't respond to any queries (including metadata fetch),
+            // but the driver will manage to continue with dummy metadata.
+            RequestReaction::forge().server_error(),
+        )))
 }
 
 pub(crate) fn drop_metadata_queries_rules() -> impl IntoIterator<Item = RequestRule> {
